requirements.txt
asgiref==3.2.10
Django==3.0
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-js-asset==1.2.2
djangorestframework==3.11.1
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
